Move payload service constants to the main dir
The constants source code contains only constants used by the main
service so it will be moved there.
