+++ pyanaconda/modules/payload/constants.py
